+++ packages/orq-rc/RELEASES.md
@@ -2939,3 +2939,13 @@ Based on:
 - [typescript v3.11.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v3.11.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.4 - packages/orq-rc
+
+## 2025-08-11 14:00:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.601.0 (2.680.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.11.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v3.11.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/3.11.0-rc.5 - packages/orq-rc